Moved all tool files into single folder
