+++ frontend/src/components/Tool/Tool_logos/AWS_logo.jpg
